Update rain chance threshold formula to scale linearly by 5% per day
- Update getMinimumPrecipProbabilityDay to return 5% * daysFromToday
- Unify day and night rain chance thresholds
- Fix unit tests in DailyForecastIconResolverTest and DailyViewLogicTest to align with new thresholds
- Fix compilation error by ensuring Int return type

+++ app/src/main/java/com/weatherwidget/widget/handlers/DailyViewLogic.kt
@@ -587,7 +587,7 @@ forecastHigh = fHigh,
             Log.d(TAG, "buildDailyRainLabel skipping today: date=$date dailyPrecip=$dailyPrecipProbability precipAmount=$precipAmountMm")
             return null
         }
-        val daysFromToday = ChronoUnit.DAYS.between(today, date)
+        val daysFromToday = ChronoUnit.DAYS.between(today, date).toInt()
         val dayMinProb = DailyForecastIconResolver.getMinimumPrecipProbabilityDay(daysFromToday)
         val nightMinProb = DailyForecastIconResolver.getMinimumPrecipProbabilityNight(daysFromToday)
         val dayPrecip = dayPrecipProbability ?: precipProbability ?: dailyPrecipProbability
@@ -626,7 +626,7 @@ forecastHigh = fHigh,
             Log.d(TAG, "buildNightRainLabel skipping null night precip: date=$date")
             return null
         }
-        val daysFromToday = ChronoUnit.DAYS.between(today, date)
+        val daysFromToday = ChronoUnit.DAYS.between(today, date).toInt()
         if (daysFromToday < 0) return null
 
         val threshold = DailyForecastIconResolver.getMinimumPrecipProbabilityNight(daysFromToday)
